+++ app/examples/using-with-mantine-contextmenu/page.tsx
@@ -43,9 +43,9 @@ export default async function UsingWithMantineContextMenuExamplePage() {
       </Txt>
       <Txt idea title="Important">
         <Box mb="sm">
-          When using {MANTINE_CONTEXTMENU_PRODUCT_NAME}, you’ll need to wrap your application in a{' '}
-          <Code>ContextMenuProvider</Code> and include the necessary CSS files <strong>in the correct order</strong>.
-          Please refer to the{' '}
+          When using <ExternalLink to={MANTINE_CONTEXTMENU_LINK}>{MANTINE_CONTEXTMENU_PRODUCT_NAME}</ExternalLink>,
+          you’ll need to wrap your application in a <Code>ContextMenuProvider</Code> and include the necessary CSS files{' '}
+          <strong>in the correct order</strong>. Please refer to the{' '}
           <ExternalLink to={`${MANTINE_CONTEXTMENU_LINK}/getting-started`}>getting started</ExternalLink> page of{' '}
           {MANTINE_CONTEXTMENU_PRODUCT_NAME} for more information.
         </Box>
@@ -72,14 +72,16 @@ export default async function UsingWithMantineContextMenuExamplePage() {
       <CodeBlock code={code['RowContextMenuExample.tsx']} />
       <PageSubtitle value="Using with scrollable tables" />
       <Txt>
-        By default, a {MANTINE_CONTEXTMENU_PRODUCT_NAME} will automatically close when the user clicks anywhere on the
-        page, hits the <Code>Escape</Code> key, scrolls the or resizes the browser window. However, when using a{' '}
-        <InternalLink to="/examples/scrollable-vs-auto-height">scrollable</InternalLink> DataTable, you’ll need to close
-        the context-menu when the user scrolls the table by passing the{' '}
+        By default, a <ExternalLink to={MANTINE_CONTEXTMENU_LINK}>{MANTINE_CONTEXTMENU_PRODUCT_NAME}</ExternalLink> will
+        automatically place an overlay over the entire page and close when the user clicks it, hits the{' '}
+        <Code>Escape</Code> key, scrolls the or resizes the browser window. However, when using a{' '}
+        <InternalLink to="/examples/scrollable-vs-auto-height">scrollable</InternalLink> DataTable, the user can still
+        scroll the table using the arrow keys, so you might need to close the context-menu when the user scrolls the
+        table by passing the{' '}
         <ExternalLink to={`${MANTINE_CONTEXTMENU_LINK}/examples/imperative-hiding/`}>
           <Code>hideContextMenu</Code> function
         </ExternalLink>{' '}
-        retuned by the <Code>useContextMenu</Code> hook to the <Code>onScroll</Code> property of the DataTable:
+        returned by the <Code>useContextMenu</Code> hook to the <Code>onScroll</Code> property of the DataTable:
       </Txt>
       <RowContextMenuOnScrollableTableExample />
       <Txt>Here’s the code for the example above:</Txt>
@@ -94,7 +96,7 @@ export default async function UsingWithMantineContextMenuExamplePage() {
       <CodeBlock code={code['CellContextMenuExample.tsx']} />
       <PageSubtitle value="Mixing row and cell context-menus" />
       <Txt>
-        You can also mix row and cell context-menus. In the example below, we’re shwing a context-menu for the{' '}
+        You can also mix row and cell context-menus. In the example below, we’re showing a context-menu for the{' '}
         <Code>name</Code> column and a different one for the row:
       </Txt>
       <RowAndCellContextMenuExample />